Add back unifrac_task_noclass.cpp, unifrac_task_impl now header-only
