+++ src/Makefile
@@ -165,12 +165,12 @@ lib$(SSU).so: tree.o biom.o biom_inmem.o biom_subsampled.o tsv.o unifrac.o cmd.o
 api.o: api.cpp api.hpp unifrac.hpp skbio_alt.hpp biom.hpp biom_inmem.hpp biom_subsampled.hpp tree.hpp tsv.hpp
 	$(CXX) $(CPPFLAGS) api.cpp -c -o api.o -fPIC
 
-unifrac_task_cpu.o: unifrac_task_noclass.cpp unifrac_task_noclass.hpp
+unifrac_task_cpu.o: unifrac_task_noclass.cpp unifrac_task_noclass.hpp unifrac_task_impl.hpp
 	$(CXX) $(CPPFLAGS) -DSUCMP_NM=su_cpu -c $< -o $@
 
-unifrac_task_acc.o: unifrac_task_noclass.cpp unifrac_task_noclass.hpp
+unifrac_task_acc.o: unifrac_task_noclass.cpp unifrac_task_noclass.hpp unifrac_task_impl.hpp
 	$(CXX) $(CPPFLAGS) $(ACCCPPFLAGS) -DSUCMP_NM=su_acc -c $< -o $@
-unifrac_task_ompgpu.o: unifrac_task_noclass.cpp unifrac_task_noclass.hpp
+unifrac_task_ompgpu.o: unifrac_task_noclass.cpp unifrac_task_noclass.hpp unifrac_task_impl.hpp
 	$(CXX) $(CPPFLAGS) $(OMPGPUCPPFLAGS) -DSUCMP_NM=su_ompgpu -c $< -o $@
 
 unifrac_cmp_cpu.o: unifrac_cmp.cpp unifrac_cmp.hpp unifrac_internal.hpp unifrac.hpp unifrac_task.hpp unifrac_task_noclass.hpp biom_interface.hpp tree.hpp
